+++ frontend/src/components/_common/errorBoundary/ListErrorFallback.tsx
@@ -29,6 +29,8 @@ export default ListErrorFallback;
 
 const S = {
   Container: styled.div`
+    width: 100%;
+    height: 50%;
     ${flexColumn}
     ${flexCenter}
     gap: 1.4rem;
